% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_phylo.R
\name{collapse_tree_all}
\alias{collapse_tree_all}
\title{collapse_tree_all}
\usage{
collapse_tree_all(tree)
}
\arguments{
\item{tree}{A annotated_phylo class object.}
}
\description{
Take a distance matrix which may contain multiple distinct samples per lesion and return a list of all possible matrices where each lesion is represented by a single sample.
}
