% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_phylo.R
\name{collapse_tree_reiter}
\alias{collapse_tree_reiter}
\title{collapse_tree_reiter}
\usage{
collapse_tree_reiter(tree, break_tie_method = "least_similar", iter.max = 500)
}
\description{
collapse_tree_reiter
}
