% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dist_similarity}
\alias{dist_similarity}
\title{dist_similarity}
\usage{
dist_similarity(
  test_dist,
  ref_dist,
  nperm = 100,
  cpus = 1,
  return_only_pval = T,
  method = "spearman"
)
}
\description{
dist_similarity
}
