% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{distance_matrix_correlation}
\alias{distance_matrix_correlation}
\title{distance_matrix_correlation}
\usage{
distance_matrix_correlation(
  test_dist_full,
  ref_dist_full,
  method,
  return_only_r = F
)
}
\description{
distance_matrix_correlation
}
