% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_phylo.R
\name{group_samples}
\alias{group_samples}
\title{group_samples}
\usage{
group_samples(
  input,
  lun = T,
  liv = T,
  per = T,
  primary_autopsy_is_distant = T,
  highlight_peritoneum_when_multi = T,
  color = F
)
}
\description{
group_samples
}
