% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_phylo.R
\name{plot.annotated_phylo}
\alias{plot.annotated_phylo}
\title{plot.annotated_phylo}
\usage{
\method{plot}{annotated_phylo}(
  tree,
  cex = 2.5,
  angle = F,
  layout = "ape",
  legend.position = "none",
  suppress_tip_labs = F,
  xpad = 0.1,
  ...
)
}
\description{
plot.annotated_phylo
}
